Remove the other use of hardcoded presets, and move the presets themselves into the test suite since they're only used there
