+++ test/javascript/RegistrationPolicy/defaultPresets.js
